+++ src/org/lockss/util/Constants.java
@@ -1,5 +1,5 @@
 /*
- * $Id: Constants.java,v 1.2 2003-03-12 02:26:46 tal Exp $
+ * $Id: Constants.java,v 1.3 2003-03-26 23:14:35 tal Exp $
  */
 
 /*
@@ -38,13 +38,13 @@ import java.util.*;
  */
 public interface Constants {
   /** The number of milliseconds in a second */
-  public static final int SECOND = 1000;
+  public static final long SECOND = 1000;
   /** The number of milliseconds in a minute */
-  public static final int MINUTE = 60 * SECOND;
+  public static final long MINUTE = 60 * SECOND;
   /** The number of milliseconds in an hour */
-  public static final int HOUR = 60 * MINUTE;
+  public static final long HOUR = 60 * MINUTE;
   /** The number of milliseconds in a day */
-  public static final int DAY = 24 * HOUR;
+  public static final long DAY = 24 * HOUR;
   /** The number of milliseconds in a week */
-  public static final int WEEK = 7 * DAY;
+  public static final long WEEK = 7 * DAY;
 }
